Update README for dev-hub parsing (#82)
In the dev-hub we are going to automatically sync the content from the cimg READMEs. 
As a part of the copy in the dev-hub, the semantic versions in the `Getting Started` header are going to be updated to the latest tag.  

https://circleci.com/developer/images/image/cimg/base
Currently, the sample config is updated with the latestversion, but the description below it mentions the text here.
As we automatically sync the content, we're simply going to update the SEMVER in the *entire header*. This change makes the content in the header make more sense.