Upgrade Dockerize to v0.8.0 (#200)
Fixes #198 

This Dockerize update is mainly being done in preparation for upcoming
arm64 image support. The current version we were using was amd64 only.
Diving a bit deeper though...

The version of Dockerize we've been using has been stuck in time. For
two reasons. 1) The original project,
https://github.com/jwilder/dockerize, is no longer maintained. So there
hasn't been new releases. 2) We were using a release from the project
stored in our S3 bucket.

We're switching to a fork that has been accepted by many on the
community. Relevant is this [GitHub
Issue](https://github.com/docker/docker.github.io/issues/13548) in the
official Docker Docs as well as this [follow up
PR](https://github.com/docker/docker.github.io/pull/13549) that actually
makes the suggestion change over to this newer project. We're switching
to this fork.

v0.8.0 was chosen for a specific reason. I don't want to jump straight
to v0.16.3 in case there are breaking changes we're not aware of. I
rather ease into that change over time. I wanted to use a version much
closer to v0.6.1 from the original project. v0.8.0 of the fork is the
closest version that was also compiled for arm64, which we need.