Prep new monthly release (#212)
This includes the work needed to publish the monthly updates but with
support for arm64 images. This work is partially manual and doesn't fit
in with the fully automated system we normally use for cimg-shared. This
is okay for now because:

1. it's temporary as work continues in cimg-shared for better arm
support
2. This image was always slightly an oddball anyway since it's the base
image.

I'm hoping this have this system more automated by the November release.